+++ tests/integration/test_integratord/test_integratord_change_inode_alert.py
@@ -4,7 +4,7 @@ copyright: Copyright (C) 2015-2022, Wazuh Inc.
            This program is free software; you can redistribute it and/or modify it under the terms of GPLv2
 
 type: integration
-brief: Integratord manages wazuh integrations with other applications such as Yara or Virustotal, by feeding
+brief: Integratord manages wazuh integrations with other applications such as Yara or Slack, by feeding
 the integrated aplications with the alerts located in alerts.json file. This test module aims to validate that
 given a specific alert, the expected response is recieved, depending if it is a valid/invalid json alert, an
 overlong alert (64kb+) or what happens when it cannot read the file because it is missing.
@@ -21,15 +21,15 @@ os_version:
     - Centos 8
     - Ubuntu Focal
 references:
-    - https://documentation.wazuh.com/current/user-manual/capabilities/virustotal-scan/integration.html
-    - https://documentation.wazuh.com/current/user-manual/reference/daemons/wazuh-integratord.htm
+    - https://documentation.wazuh.com/current/user-manual/manager/manual-integration.html#slack
+    - https://documentation.wazuh.com/current/user-manual/reference/daemons/wazuh-integratord.html
 pytest_args:
     - tier:
         0: Only level 0 tests are performed, they check basic functionalities and are quick to perform.
         1: Only level 1 tests are performed, they check functionalities of medium complexity.
         2: Only level 2 tests are performed, they check advanced functionalities and are slow to perform.
 tags:
-    - virustotal
+    - slack
 '''
 import os
 import time
@@ -53,18 +53,19 @@ CONFIGURATIONS_PATH = os.path.join(TEST_DATA_PATH, 'configuration_template')
 TEST_CASES_PATH = os.path.join(TEST_DATA_PATH, 'test_cases')
 
 # Configuration and cases data
-configurations_path = os.path.join(CONFIGURATIONS_PATH, 'config_integratord_read_json_alerts.yaml')
+configurations_path = os.path.join(CONFIGURATIONS_PATH, 'configuration_integratord_read_json_alerts.yaml')
 cases_path = os.path.join(TEST_CASES_PATH, 'cases_integratord_change_inode_alert.yaml')
 
 # Configurations
 configuration_parameters, configuration_metadata, case_ids = get_test_cases_data(cases_path)
-configuration_parameters[0]['API_KEY'] = global_parameters.integration_api_key
+configuration_parameters[0]['WEBHOOK_URL'] = global_parameters.slack_webhook_url
 configurations = load_configuration_template(configurations_path, configuration_parameters,
                                              configuration_metadata)
-local_internal_options = {'integrator.debug': '2'}
+local_internal_options = {'integrator.debug': '2', 'analysisd.debug': '1'}
 
 # Variables
 TEMP_FILE_PATH = os.path.join(WAZUH_PATH, 'logs/alerts/alerts.json.tmp')
+REQUIRED_DAEMONS = integrator.REQUIRED_DAEMONS
 
 
 # Tests
@@ -78,7 +79,7 @@ def test_integratord_change_json_inode(configuration, metadata, set_wazuh_config
     description: Check that when a given alert is inserted into alerts.json, integratord works as expected.
     wazuh_min_version: 4.3.5
     tier: 1
-   parameters:
+    parameters:
         - configuration:
             type: dict
             brief: Configuration loaded from `configuration_template`.
@@ -96,17 +97,19 @@ def test_integratord_change_json_inode(configuration, metadata, set_wazuh_config
             brief: Configure the local internal options file.
         - restart_wazuh_function:
             type: fixture
-            brief: Restart wazuh-modulesd daemon before starting a test, and stop it after finishing.
+            brief: Restart a list of daemons (defined in REQUIRED_DAEMONS variable) and stop them after finishing.
         - wait_for_start_module:
             type: fixture
             brief: Detect the start of the Integratord module in the ossec.log
     assertions:
         - Verify the expected response with for a given alert is recieved
     input_description:
-        - The `config_integratord_read_json_alerts.yaml` file provides the module configuration for this test.
+        - The `configuration_integratord_read_json_alerts.yaml` file provides the module configuration for this test.
         - The `cases_integratord_read_json_alerts` file provides the test cases.
     expected_output:
-        - r'.*Sending FIM event: (.+)$' ('added', 'modified' and 'deleted' events)
+        - r'.*wazuh-integratord.*DEBUG: sending new alert'
+        - r'.*wazuh-integratord.*DEBUG: jqueue_next.*Alert file inode changed.*'
+        - r'.*wazuh-integratord.*Processing alert.*'
 
     '''
     wazuh_monitor = FileMonitor(LOG_FILE_PATH)
@@ -114,7 +117,7 @@ def test_integratord_change_json_inode(configuration, metadata, set_wazuh_config
     # Insert Alerts
     run_local_command_returning_output(command)
 
-    # Get that alert is read
+    # Check that the alert was read
     check_integratord_event(file_monitor=wazuh_monitor, timeout=global_parameters.default_timeout,
                             callback=callback_generator(integrator.CB_INTEGRATORD_SENDING_ALERT),
                             error_message=integrator.ERR_MSG_SENDING_ALERT_NOT_FOUND,
@@ -125,16 +128,20 @@ def test_integratord_change_json_inode(configuration, metadata, set_wazuh_config
     remove_file(ALERT_FILE_PATH)
     copy(TEMP_FILE_PATH, ALERT_FILE_PATH)
 
-    # Wait for Inode change to be detected and insert new alert
-    time.sleep(3)
-    run_local_command_returning_output(command)
+    # Wait for Inode change to be detected
+    # The `integratord` library tries to read alerts from the file every 1 second. So, the test waits 1 second + 1
+    # until the file is reloaded.
+    time.sleep(integrator.TIME_TO_DETECT_FILE)
 
     # Monitor Inode Changed
-    check_integratord_event(file_monitor=wazuh_monitor, timeout=global_parameters.default_timeout * 2,
+    check_integratord_event(file_monitor=wazuh_monitor, timeout=global_parameters.default_timeout,
                             callback=callback_generator(integrator.CB_ALERTS_FILE_INODE_CHANGED),
                             error_message=integrator.ERR_MSG_ALERT_INODE_CHANGED_NOT_FOUND)
+
+    # Insert a new alert
+    run_local_command_returning_output(command)
 
     # Read Response in ossec.log
     check_integratord_event(file_monitor=wazuh_monitor, timeout=global_parameters.default_timeout,
                             callback=callback_generator(integrator.CB_PROCESSING_ALERT),
-                            error_message=integrator.ERR_MSG_VIRUSTOTAL_ALERT_NOT_DETECTED)
+                            error_message=integrator.ERR_MSG_SLACK_ALERT_NOT_DETECTED)